Extract mission list gui component
