+++ src/app/gates/detail/detail.component.html
@@ -39,16 +39,7 @@
         </div>
         <div *ngIf="data.missions.length">
             <h3>Missions</h3>
-            <ul class="mission-list" *ngIf="$missions | async; let missions">
-                <ng-container *ngFor="let mission of missions">
-                    <li *luxVar="mission.$table | async; let m" [style.order]="mission.$sortOrder | async">
-                        <lux-mission [id]="mission.id" [isMission]="m ? m.isMission : false"
-                            [iconID]="mission.$missionIcon | async" [title]="(mission.$loc | async)?.name"
-                            [tooltip]="((m && !m.isMission) ? 'Achievement' : 'Mission') + ' #' + mission.id"></lux-mission>
-                        <!--{{mission.$table | async | json}}-->
-                    </li>
-                </ng-container>
-            </ul>
+            <lux-mission-list [$missionIds]="$missionIds"></lux-mission-list>
         </div>
         <div *ngIf="data.mission_tasks.length">
             <h3>Mission Tasks</h3>
